data-generator/requirements.txt
kafka-python==2.0.2
pandas==2.0.3
faker==19.6.2
python-dotenv==1.0.0